* 'master' of https://github.com/AY2021S1-CS2103T-F13-3/tp: (89 commits)
  Update UserGuide.md
  Change XYChart.Data to XYChart.Data<>
  Change date format in charts to dd-mm-yy
  Update UserGuide.md
  Make charts available even when height/weight is present without the other
  Fix scroll bug for charts
  Update UI
  Update add command constraints message
  Update PatientChartCard.java
  Fix bug for loading a patient without any past records
  Update createAccount so that admin cannot be used as username
  Fix bug where user can still invoke access command from a patient profile page
  Readjust graph locations
  Update UserGuide.md
  Update Mcr constructor to convert mcr to all caps
  Update login admin account details in UG
  Remove print line used for testing
  Fix bug where accounts.json is not created properly
  Update add feature
  Display index for medical notes
  ...

+++ src/main/java/seedu/medibook/logic/commands/DeleteNoteCommand.java
@@ -31,6 +31,8 @@ public class DeleteNoteCommand extends Command {
     public static final String MESSAGE_DELETE_NOTE_ON_LIST = "You can only delete a medical note from a patient "
             + "when you are viewing his/her patient profile. Access the patient profile before "
             + "deleting a medical note.";
+    public static final String MESSAGE_USER_CANNOT_DELETE = "Current user cannot delete medical notes";
+    public static final String MESSAGE_CANNOT_DELETE_OTHER_DOCTOR_NOTES = "Can't delete other doctor's medical notes";
 
     private final Logger logger = LogsCenter.getLogger(DeleteNoteCommand.class);
 
@@ -66,9 +68,17 @@ public class DeleteNoteCommand extends Command {
             throw new CommandException(Messages.MESSAGE_INVALID_NOTE_DISPLAYED_INDEX);
         }
 
+        if (model.getActiveUser().isEmpty()) {
+            throw new CommandException(MESSAGE_USER_CANNOT_DELETE);
+        }
+
         MedicalNote noteToDelete = medicalNoteList.getMedicalNoteAtIndex(zeroBasedIndex);
 
-        medicalNoteList.deleteMedicalNoteAtIndex(zeroBasedIndex);
+        if (noteToDelete.getDoctorMcr().equals(model.getActiveUser().get().getMcr())) {
+            medicalNoteList.deleteMedicalNoteAtIndex(zeroBasedIndex);
+        } else {
+            throw new CommandException(MESSAGE_CANNOT_DELETE_OTHER_DOCTOR_NOTES);
+        }
 
         logger.info("----------------[PATIENT AND ORDER OF MEDICAL NOTES:][" + displayedPatient + "\n"
                 + displayedPatient.getMedicalNoteList() + "]");
